requirements.txt
Django==1.11.20
djangorestframework==3.9.1
psycopg2==2.5.4
pillow==5.4.1
django-filter==1.1.0
celery==4.2.2
django-debug-toolbar==1.11
django-dynamic-fixture==2.0.0
django-debug-panel==0.8.3
coverage==4.5.3